[IMP] Implements lot selection on delivery orders.
Adds the ability to select specific lot/serial numbers when validating delivery orders.

Introduces a wizard that allows users to choose the lots to be used for a stock move, enhancing control over lot allocation during deliveries.
This is enabled via a new flag on picking types, allowing to select existing lots.